Move component UntrackedRateCard from 'components/cards' to 'components/index/CardsMonitoring'
